+++ Gravitas.Platform.Web/ViewModel/OpRoutine/LabolatoryIn/01_IdleVm.cs
@@ -1,10 +1,11 @@
-﻿namespace Gravitas.Platform.Web.ViewModel {
-
-	public static partial class LaboratoryInVms {
-
-		public class IdleVm {
-			public long NodeId { get; set; }
-			public bool IsBusy { get; set; }
-		}
-	}
+﻿namespace Gravitas.Platform.Web.ViewModel
+{
+    public static partial class LaboratoryInVms
+    {
+        public class IdleVm
+        {
+            public int NodeId { get; set; }
+            public bool IsBusy { get; set; }
+        }
+    }
 }
