+++ Gravitas.Platform.Web/ViewModel/OpRoutine/UnloadPointType1/IdleVm.cs
@@ -1,10 +1,11 @@
-﻿namespace Gravitas.Platform.Web.ViewModel {
-
-	public static partial class UnloadPointType1Vms {
-
-		public class IdleVm {
-			public long NodeId { get; set; }
-			public UnloadPointTicketContainerItemVm BindedTruck { get; set; }
-		}
-	}
+﻿namespace Gravitas.Platform.Web.ViewModel
+{
+    public static partial class UnloadPointType1Vms
+    {
+        public class IdleVm
+        {
+            public int NodeId { get; set; }
+            public UnloadPointTicketContainerItemVm BindedTruck { get; set; }
+        }
+    }
 }
